refactor: make a separate component for the homepage logo to be reusable
