+++ src/Components/Common/index.js
@@ -1,3 +1,4 @@
 export * from './Button';
 export * from './Card';
 export * from './CardSection';
+export * from './HomeImage';
