+++ src/Components/CreateJoin.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import { View, Text, Image } from 'react-native';
-import { Card, CardSection, Button } from './Common';
+import { Card, CardSection, Button, HomeImage } from './Common';
 
 export default class CreateJoin extends Component {
 
@@ -15,12 +14,7 @@ export default class CreateJoin extends Component {
   render () {
     return (
         <Card>
-          <CardSection>
-            <Image
-              style={styles.imageStyle}
-              source={{uri: imageUrl}}
-            />
-          </CardSection>
+          <HomeImage />
 
           <CardSection>
             <Button onPress={this.onCreateClick.bind(this)}>
@@ -37,14 +31,3 @@ export default class CreateJoin extends Component {
     );
   }
 }
-
-const imageUrl = 'https://lh3.googleusercontent.com/Je1l2lhriY7bJ6EfoMWW48FI2-9WSV7DM9i_xCYJXYMI1BtImqs-aRVz-thTOMccCh5qI0NBrMnv4ybudwlj=w1523-h904';
-const styles = {
-  imageStyle: {
-    height: 400,
-    flex: 1,
-    width: null,
-    marginTop: 30,
-    marginBottom: 30
-  }
-};
